Move failure window type and methods to own file
